chore: Swagger UI의 응답과 요청의 Example Values 설정 (#29)
* Swagger UI 엔드포인트 제목 작성

* 카테고리 생성 엔드포인트와 요청 응답 DTO에 대한 Swagger 설명 추가 & 응답 레코드를 도메인 레코드와 분리

* 카테고리 트리 조회 엔드포인트 스키마 작성 및 이름 수정

* 내 댓글 목록 조회 엔드포인트 스키마 작성

* 프로젝트에 댓글 작성 엔드포인트 스키마 작성

* 프로젝트 댓글 엔드포인트 스키마 작성

* 결제 요청 엔드포인트 스키마 작성

* 결제 상세 조회 엔드포인트 스키마 작성 & 명칭 변경

* 결제 취소 엔드포인트 스키마 작성

* 참여한 후원의 배송정보 엔드포인트 스키마 작성

* 관리자용 후원 엔드포인트 스키마 작성

* 창작자용 후원 엔드포인트 스키마 작성

* 사용자용 후원 엔드포인트 스키마 작성

* 관리자용 프로젝트 엔드포인트 스키마 작성

* 창작자용 프로젝트 엔드포인트 스키마 작성

* 사용자용 프로젝트 엔드포인트 스키마 작성

* 창작자용 리워드 엔드포인트 스키마 작성

* 사용자용 리워드 엔드포인트 스키마 작성

* 유저 엔드포인트 스키마 작성

* 유저 배송지 엔드포인트 스키마 작성

* Swagger UI 문서에 표시될 컨트롤러의 이름 및 정렬 순서 변경

* AuthController의 엔드포인트가 Swagger UI 문서에 표시될 example value 설정

* CategoryController의 엔드포인트가 Swagger UI 문서에 표시될 example value 설정

* CommentController의 엔드포인트가 Swagger UI 문서에 표시될 example value 설정

* PaymentController의 엔드포인트가 Swagger UI 문서에 표시될 example value 설정 & Payment와 Pledge의 amount 데이터 타입을 Long에서 BigDecimal로 변경

* 중간 패키지 오타 수정

* FulfillmentStatus 값 수정 & UserPledgeCreateRequest의 변수명 수정 & PledgeController의 엔드포인트가 Swagger UI에 보여줄 Example Value 수정

* PledgeAddressController의 엔드포인트가 Swagger UI에 보여줄 Example Value 수정

* ProjectController의 엔드포인트가 Swagger UI에 보여줄 Example Value 수정

* RewardController의 엔드포인트가 Swagger UI에 보여줄 Example Value 수정

* UserController의 엔드포인트가 Swagger UI에 보여줄 Example Value 수정

* UserAddressController의 엔드포인트가 Swagger UI에 보여줄 Example Value 수정

+++ src/main/java/io/github/crowdfund/domain/pledge/Pledge.java
@@ -3,6 +3,7 @@ package io.github.crowdfund.domain.pledge;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -23,7 +24,7 @@ public record Pledge(
     Long userId,
     Long projectId,
     Long rewardId,
-    Long amount,
+    BigDecimal amount,
     FulfillmentStatus fulfillmentStatus,
     LocalDateTime fulfilledAt,
     LocalDateTime createdAt
